refactor(api): move roles to modules/ and share the permission matrix schema
